// web/src/gen/pb/api/v1/data.ts
export interface ListTabSummariesResponse {
    /**
     * List of tab summaries.
     *
     * @generated from protobuf field: repeated testgrid.api.v1.TabSummary tab_summaries = 1
     */
    tabSummaries: TabSummary[];
}
// ...
export interface TabSummary {
    /**
     * The name of the dashboard this tab belongs to.
     *
     * @generated from protobuf field: string dashboard_name = 1
     */
    dashboardName: string;
    /**
     * The name of the tab.
     *
     * @generated from protobuf field: string tab_name = 2
     */
    tabName: string;
    /**
     * Overall status for the dashboard tab (e.g. PASSING, FAILING, etc.)
     *
     * @generated from protobuf field: string overall_status = 3
     */
    overallStatus: string;
    /**
     * Summary of the status for this dashboard tab.
     *
     * @generated from protobuf field: string detailed_status_message = 4
     */
    detailedStatusMessage: string;
    /**
     * Timestamp at which tests last ran.
     *
     * @generated from protobuf field: google.protobuf.Timestamp last_run_timestamp = 5
     */
    lastRunTimestamp?: Timestamp;
    /**
     * Timestamp at which the test group was last updated.
     *
     * @generated from protobuf field: google.protobuf.Timestamp last_update_timestamp = 6
     */
    lastUpdateTimestamp?: Timestamp;
    /**
     * The ID for the latest passing build.
     *
     * @generated from protobuf field: string latest_passing_build = 7
     */
    latestPassingBuild: string;
    /**
     * Summarized info on the failing tests.
     * In this case, any test which raised an alert and did not suppress it is considered failing.
     *
     * @generated from protobuf field: testgrid.api.v1.FailuresSummary failures_summary = 8
     */
    failuresSummary?: FailuresSummary;
    /**
     * Summarized info on the tab's healthiness.
     *
     * @generated from protobuf field: testgrid.api.v1.HealthinessSummary healthiness_summary = 9
     */
    healthinessSummary?: HealthinessSummary;
}
/**
 * Summarized representation of data from failing test summaries.
 * Will be rendered in failures summary component within tab summary.
 *
 * @generated from protobuf message testgrid.api.v1.FailuresSummary
 */
export interface FailuresSummary {
    /**
     * Top failing tests by fail count.
     *
     * @generated from protobuf field: repeated testgrid.api.v1.FailingTestInfo top_failing_tests = 1
     */
    topFailingTests: FailingTestInfo[];
    /**
     * Aggregated stats across all failing tests.
     *
     * @generated from protobuf field: testgrid.api.v1.FailureStats failure_stats = 2
     */
    failureStats?: FailureStats;
}
/**
 * Subset of data from FailingTestSummary defined in summary.proto.
 *
 * @generated from protobuf message testgrid.api.v1.FailingTestInfo
 */
export interface FailingTestInfo {
    /**
     * Name of the failing test.
     *
     * @generated from protobuf field: string display_name = 1
     */
    displayName: string;
    /**
     * number of times the test has failed.
     *
     * @generated from protobuf field: int32 fail_count = 2
     */
    failCount: number;
    /**
     * Timestamp for the last cycle in which the test passed.
     *
     * @generated from protobuf field: google.protobuf.Timestamp pass_timestamp = 3
     */
    passTimestamp?: Timestamp;
    /**
     * Timestamp for the first cycle in which the test failed.
     *
     * @generated from protobuf field: google.protobuf.Timestamp fail_timestamp = 4
     */
    failTimestamp?: Timestamp;
}
/**
 * Aggregate stats across all failing tests.
 *
 * @generated from protobuf message testgrid.api.v1.FailureStats
 */
export interface FailureStats {
    /**
     * Number of failing tests for the tab.
     *
     * @generated from protobuf field: int32 num_failing_tests = 1
     */
    numFailingTests: number;
}
/**
 * Summarized representation of data from tab's HealthinessInfo.
 * Will be rendered in healthiness summary component within tab summary.
 *
 * @generated from protobuf message testgrid.api.v1.HealthinessSummary
 */
export interface HealthinessSummary {
    /**
     * Top flaky tests (with flakiness > 0) by the current flakiness %.
     *
     * @generated from protobuf field: repeated testgrid.api.v1.FlakyTestInfo top_flaky_tests = 1
     */
    topFlakyTests: FlakyTestInfo[];
    /**
     * Aggregated healthiness stats for the tab.
     *
     * @generated from protobuf field: testgrid.api.v1.HealthinessStats healthiness_stats = 2
     */
    healthinessStats?: HealthinessStats;
}
/**
 * Subset of data from HealthinessInfo.TestInfo defined in summary.proto.
 *
 * @generated from protobuf message testgrid.api.v1.FlakyTestInfo
 */
export interface FlakyTestInfo {
    /**
     * Name of the flaky test.
     *
     * @generated from protobuf field: string display_name = 1
     */
    displayName: string;
    /**
     * The flakiness of the test, in % measured out of 100
     *
     * @generated from protobuf field: float flakiness = 2
     */
    flakiness: number;
    /**
     * The change of flakiness based on the last interval's flakiness
     * The interval is set by each tab's config, with a default of 7 days.
     *
     * @generated from protobuf field: testgrid.summary.TestInfo.Trend change = 3
     */
    change: TestInfo_Trend;
}
/**
 * Aggregated healthiness stats across the tab.
 *
 * @generated from protobuf message testgrid.api.v1.HealthinessStats
 */
export interface HealthinessStats {
    /**
     * The start of the time frame that the analysis was run for.
     *
     * @generated from protobuf field: google.protobuf.Timestamp start = 1
     */
    start?: Timestamp;
    /**
     * The end of the time frame that the analysis was run for.
     *
     * @generated from protobuf field: google.protobuf.Timestamp end = 2
     */
    end?: Timestamp;
    /**
     * Number of flaky tests in the tab.
     *
     * @generated from protobuf field: int32 num_flaky_tests = 3
     */
    numFlakyTests: number;
    /**
     * Average flakiness for the current analysis interval.
     *
     * @generated from protobuf field: float average_flakiness = 4
     */
    averageFlakiness: number;
    /**
     * Average flakiness for the previous analysis interval.
     *
     * @generated from protobuf field: float previous_flakiness = 5
     */
    previousFlakiness: number;
}
// ...
class ListTabSummariesResponse$Type extends MessageType<ListTabSummariesResponse> {
    constructor() {
        super("testgrid.api.v1.ListTabSummariesResponse", [
            { no: 1, name: "tab_summaries", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => TabSummary }
        ]);
    }
    create(value?: PartialMessage<ListTabSummariesResponse>): ListTabSummariesResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.tabSummaries = [];
        if (value !== undefined)
            reflectionMergePartial<ListTabSummariesResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListTabSummariesResponse): ListTabSummariesResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated testgrid.api.v1.TabSummary tab_summaries */ 1:
                    message.tabSummaries.push(TabSummary.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListTabSummariesResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated testgrid.api.v1.TabSummary tab_summaries = 1; */
        for (let i = 0; i < message.tabSummaries.length; i++)
            TabSummary.internalBinaryWrite(message.tabSummaries[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message testgrid.api.v1.ListTabSummariesResponse
 */
export const ListTabSummariesResponse = new ListTabSummariesResponse$Type();
// ...
export const TabSummary = new TabSummary$Type();
// @generated message type with reflection information, may provide speed optimized methods
class FailuresSummary$Type extends MessageType<FailuresSummary> {
    constructor() {
        super("testgrid.api.v1.FailuresSummary", [
            { no: 1, name: "top_failing_tests", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => FailingTestInfo },
            { no: 2, name: "failure_stats", kind: "message", T: () => FailureStats }
        ]);
    }
    create(value?: PartialMessage<FailuresSummary>): FailuresSummary {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.topFailingTests = [];
        if (value !== undefined)
            reflectionMergePartial<FailuresSummary>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: FailuresSummary): FailuresSummary {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated testgrid.api.v1.FailingTestInfo top_failing_tests */ 1:
                    message.topFailingTests.push(FailingTestInfo.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* testgrid.api.v1.FailureStats failure_stats */ 2:
                    message.failureStats = FailureStats.internalBinaryRead(reader, reader.uint32(), options, message.failureStats);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: FailuresSummary, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated testgrid.api.v1.FailingTestInfo top_failing_tests = 1; */
        for (let i = 0; i < message.topFailingTests.length; i++)
            FailingTestInfo.internalBinaryWrite(message.topFailingTests[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* testgrid.api.v1.FailureStats failure_stats = 2; */
        if (message.failureStats)
            FailureStats.internalBinaryWrite(message.failureStats, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message testgrid.api.v1.FailuresSummary
 */
export const FailuresSummary = new FailuresSummary$Type();
// ...
export const FailingTestInfo = new FailingTestInfo$Type();
// ...
export const FailureStats = new FailureStats$Type();
// ...
export const HealthinessSummary = new HealthinessSummary$Type();
// ...
export const FlakyTestInfo = new FlakyTestInfo$Type();
// ...
export const HealthinessStats = new HealthinessStats$Type();